third_party/blink/web_tests/external/wpt/focus/focus-restoration-in-different-site-iframes-expected.txt

This is a testharness.js-based test.
[FAIL] Check result
  assert_equals: Check log expected "outerlog:log:willfocusinput,bodyfocus,inputfocus,didfocusinput,inputblur,bodyblur,bodyfocus,inputfocus," but got "outerlog:log:willfocusinput,didfocusinput,"
Harness: the test ran to completion.
